requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
gTTS==2.3.1
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
multidict==6.0.4
openai==0.27.4
OpenAIAuth==0.3.6
playsound==1.3.0
prompt-toolkit==3.0.38
pycairo==1.23.0
pygame==2.3.0
PySocks==1.7.1
python-dotenv==1.0.0
regex==2023.3.23
requests==2.28.2
revChatGPT==4.2.5
sniffio==1.3.0
socksio==1.0.0
SpeechRecognition==3.10.0
tiktoken==0.3.3
tqdm==4.65.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.9.1
